Split fullscreen composer internals
